# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel blog Api Documentation",
        "description": "You will get all blog api",
        "contact": {
            "email": "ankush0094@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/laravel_blog_api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/v1/post": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Post listing",
                "description": "get all post",
                "operationId": "Listing",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        },
                        "example": "20"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/post/add": {
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "Add",
                "description": "Add new post",
                "operationId": "Add",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "tags[0]": {
                                        "type": "string"
                                    },
                                    "tags[1]": {
                                        "type": "string"
                                    },
                                    "media": {
                                        "type": "file"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "title": "demo title",
                                    "content": "content title",
                                    "tags[0]": "tag1",
                                    "tags[1]": "tag2"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/post/edit/{id}": {
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "Update",
                "description": "UPdate post",
                "operationId": "Update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "post id parameter",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "tags[0]": {
                                        "type": "string"
                                    },
                                    "tags[1]": {
                                        "type": "string"
                                    },
                                    "media": {
                                        "type": "file"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "title": "demo title",
                                    "content": "content title",
                                    "tags[0]": "tag1",
                                    "tags[1]": "tag2"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/post/delete/{id}": {
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "Delete",
                "description": "Delete post",
                "operationId": "Delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "post id parameter",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "422": {
                        "description": "validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/tag/search": {
            "get": {
                "tags": [
                    "Tags"
                ],
                "summary": "Tag search",
                "description": "Search tags",
                "operationId": "Tags",
                "parameters": [
                    {
                        "name": "term",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "php"
                    }
                ],
                "responses": {
                    "422": {
                        "description": "validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}